requirements.txt
streamlit
nltk
ipynb
langchain
ipykernel
python-dotenv
langchain-community
pypdf
bs4
arxiv
PyMuPDF
wikipedia
langchain-text-splitters
langchain-openai
chromadb
langchain-ollama
sentence-transformers
langchain-huggingface
faiss-cpu
langchain_chroma
langsmith
beautifulsoup4
langchain-groq
langchain-core
fastapi
uvicorn
langserve
sse_starlette
pydantic
sentence-transformers
langchain_chroma
ddgs
duckduckgo-search
SQLAlchemy
mysql-connector-python
pymysql
langgraph>=1.0.0
langchain-openai>=1.0.0
langchain-anthropic>=1.0.0
langchain-community
jupyter>=1.0.0
langgraph-cli[inmem]>=0.4.0
validators==0.28.1
youtube-transcript-api==1.2.0
unstructured
pytube
langchain_huggingface
huggingface_hub
gradio